there is weird building break for release and debug, not sure why but it is after adding picoscrew

+++ Chimera/Source/GeneralObjects/Queues.h
@@ -7,7 +7,7 @@
 
 // info for the picture in the experiment run in which the pic is taken
 struct picureStatus {
-	unsigned picNum;
+	unsigned long long picNum;
 	unsigned repNum;
 	unsigned varNum;
 };
